+++ src/main/java/frc/robot/commands/Cargo_Score/cargoScoreClawClose.java
@@ -26,14 +26,13 @@ public class cargoScoreClawClose extends Command {
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
-    
+
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.CargoScore.clawClose();
-
+    // Robot.CargoScore.clawClose();
   }
 
   // Make this return true when this Command no longer needs to run execute()
@@ -45,7 +44,7 @@ public class cargoScoreClawClose extends Command {
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    
+
   }
 
   // Called when another command which requires one or more of the same
@@ -54,4 +53,5 @@ public class cargoScoreClawClose extends Command {
   protected void interrupted() {
     this.end();
   }
+
 }
