--- src/main/java/frc/robot/subsystems/cargoScore.java
@@ -1,25 +1,0 @@
-package frc.robot.subsystems;
-
-import frc.robot.commands.*;
-
-import edu.wpi.first.wpilibj.command.Subsystem;
-
-/**
- * 
- */
-public class CargoScore extends Subsystem {
-
-
-
-  public CargoScore() {
-
-  }
-
-  // Set the default command for a subsystem here.
-
-  @Override
-  public void initDefaultCommand() {
-    setDefaultCommand(new doNothing());
-  }
-
-}
